Moved oss licenses code to separate `oss_licenses` module, cleaned dependencies
