+++ app/build.gradle
@@ -135,6 +135,7 @@ android {
         implementation project(':aws')
         implementation project(':push')
         implementation project(':automatic_backups')
+        implementation project(':oss_licenses')
 
 
         implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
@@ -199,8 +200,6 @@ android {
         // Dagger
         kapt "com.google.dagger:dagger-compiler:$DAGGER_VERSION"
         kapt "com.google.dagger:dagger-android-processor:$DAGGER_VERSION"
-
-        implementation "com.google.android.gms:play-services-oss-licenses:$GOOGLE_OSS_VERSION"
 
         // Unit Tests
         testImplementation "org.robolectric:robolectric:$ROBOLECTRIC_VERSION"
