+++ app/src/flossFlavor/java/co.smartreceipts/android/di/FlavorModule.java
@@ -7,8 +7,6 @@ import co.smartreceipts.analytics.AnalyticsProvider;
 import co.smartreceipts.android.ExtraInitializer;
 import co.smartreceipts.android.ExtraInitializerFlossImpl;
 import co.smartreceipts.android.analytics.AnalyticsManager;
-import co.smartreceipts.android.licenses.LicensesNavigator;
-import co.smartreceipts.android.licenses.NoOpLicensesNavigator;
 import co.smartreceipts.android.ocr.NoOpOcrManager;
 import co.smartreceipts.android.ocr.OcrManager;
 import co.smartreceipts.android.purchases.wallet.DefaultPurchaseWallet;
@@ -66,7 +64,4 @@ public abstract class FlavorModule {
     @ApplicationScope
     public abstract GoogleDriveTableManager provideGoogleDriveTableManager(NoOpGoogleDriveTableManager driveTableManager);
 
-    @Binds
-    @ApplicationScope
-    public abstract LicensesNavigator provideLicensesNavigator(NoOpLicensesNavigator licensesNavigator);
 }
