+++ app/src/main/java/co/smartreceipts/android/SmartReceiptsApplication.java
@@ -14,7 +14,8 @@ import org.jetbrains.annotations.NotNull;
 import javax.inject.Inject;
 
 import co.smartreceipts.analytics.Analytics;
-import co.smartreceipts.android.analytics.crash.CrashReporter;
+import co.smartreceipts.analytics.crash.CrashReporter;
+import co.smartreceipts.analytics.log.Logger;
 import co.smartreceipts.android.date.DateFormatter;
 import co.smartreceipts.android.di.AppComponent;
 import co.smartreceipts.android.di.BaseAppModule;
@@ -28,6 +29,7 @@ import co.smartreceipts.android.purchases.PurchaseManager;
 import co.smartreceipts.android.rating.data.AppRatingPreferencesStorage;
 import co.smartreceipts.android.receipts.ordering.ReceiptsOrderer;
 import co.smartreceipts.android.settings.UserPreferenceManager;
+import co.smartreceipts.android.settings.catalog.UserPreference;
 import co.smartreceipts.android.sync.cleanup.MarkedForDeletionCleaner;
 import co.smartreceipts.android.utils.StrictModeConfiguration;
 import co.smartreceipts.android.utils.WBUncaughtExceptionHandler;
@@ -37,7 +39,6 @@ import co.smartreceipts.android.utils.rx.DefaultRxErrorHandler;
 import co.smartreceipts.android.versioning.AppVersionManager;
 import co.smartreceipts.aws.cognito.CognitoManager;
 import co.smartreceipts.core.identity.IdentityManager;
-import co.smartreceipts.analytics.log.Logger;
 import co.smartreceipts.push.PushManager;
 import co.smartreceipts.push.PushManagerProvider;
 import dagger.android.AndroidInjector;
