--- app/src/main/java/co/smartreceipts/android/licenses/LicensesNavigator.kt
@@ -1,9 +1,0 @@
-package co.smartreceipts.android.licenses
-
-import android.content.Context
-import android.content.Intent
-
-interface LicensesNavigator {
-
-    fun getLicensesActivityIntent(context: Context): Intent?
-}
