+++ app/src/plusFlavor/java/co/smartreceipts/android/di/FlavorModule.java
@@ -10,8 +10,6 @@ import co.smartreceipts.analytics.AnalyticsProvider;
 import co.smartreceipts.android.ExtraInitializer;
 import co.smartreceipts.android.ExtraInitializerPlusImpl;
 import co.smartreceipts.android.analytics.AnalyticsManager;
-import co.smartreceipts.android.licenses.LicensesNavigator;
-import co.smartreceipts.android.licenses.LicensesNavigatorImpl;
 import co.smartreceipts.android.ocr.OcrManager;
 import co.smartreceipts.android.ocr.OcrManagerImpl;
 import co.smartreceipts.android.purchases.wallet.PlusPurchaseWallet;
@@ -69,7 +67,4 @@ public abstract class FlavorModule {
     @ApplicationScope
     public abstract GoogleDriveTableManager provideGoogleDriveTableManager(GoogleDriveTableManagerImpl driveTableManager);
 
-    @Binds
-    @ApplicationScope
-    public abstract LicensesNavigator provideLicensesNavigator(LicensesNavigatorImpl licensesNavigator);
 }
