+++ oss_licenses/src/main/java/co/smartreceipts/oss_licenses/LicensesNavigatorInterface.kt
@@ -1,0 +1,9 @@
+package co.smartreceipts.oss_licenses
+
+import android.content.Context
+import android.content.Intent
+
+interface LicensesNavigatorInterface {
+
+    fun getLicensesActivityIntent(context: Context, ossActivityTitleId: Int): Intent?
+}
