+++ settings.gradle
@@ -1,2 +1,2 @@
-include ':app', ':aws', ':core', ':push', ':analytics', ':automatic_backups'
+include ':app', ':aws', ':core', ':push', ':analytics', ':automatic_backups', ':oss_licenses'
 include ':wbMiniLibrary'
